+++ src/plugins/data/common/es_query/filters/get_filter_field.test.ts
@@ -21,7 +21,7 @@ import { buildPhraseFilter } from './phrase_filter';
 import { buildQueryFilter } from './query_string_filter';
 import { getFilterField } from './get_filter_field';
 import { IIndexPattern } from '../../index_patterns';
-import { fields } from '../../index_patterns/fields/fields.mocks.ts';
+import { fields } from '../../index_patterns/fields/fields.mocks';
 
 describe('getFilterField', function () {
   const indexPattern: IIndexPattern = ({
