+++ src/plugins/data/common/index_patterns/fields/fields.mocks.ts
