+++ src/plugins/data/common/index_patterns/index_patterns/_fields_fetcher.ts
