+++ src/plugins/data/common/index_patterns/index_patterns/_pattern_cache.ts
