+++ src/plugins/data/common/index_patterns/index_patterns/flatten_hit.ts
