+++ src/plugins/data/common/index_patterns/index_patterns/format_hit.ts
