+++ src/plugins/data/common/index_patterns/index_patterns/index_patterns.test.ts
