+++ src/plugins/data/common/index_patterns/index_patterns/index_patterns_api_client.test.mock.ts
