+++ src/plugins/data/common/index_patterns/lib/errors.ts
