+++ src/plugins/data/common/index_patterns/lib/get_from_saved_object.ts
