+++ src/plugins/data/common/index_patterns/lib/get_title.ts
