+++ src/plugins/data/common/index_patterns/lib/index.ts
