+++ src/plugins/data/common/index_patterns/lib/is_default.ts
