+++ src/plugins/data/common/index_patterns/lib/types.ts
