+++ src/plugins/data/common/index_patterns/lib/validate_index_pattern.test.ts
