+++ src/plugins/data/public/index_patterns/index.ts
@@ -25,12 +25,22 @@ export {
   validateIndexPattern,
   getFromSavedObject,
   isDefault,
-} from './lib';
+} from '../../common/index_patterns/lib';
 export { flattenHitWrapper, formatHitProvider } from './index_patterns';
 
-export { getIndexPatternFieldListCreator, Field, IIndexPatternFieldList } from './fields';
+export {
+  getIndexPatternFieldListCreator,
+  Field,
+  IIndexPatternFieldList,
+  IndexPatternsService,
+  IndexPatternsContract,
+  IndexPattern,
+  TypeMeta,
+  AggregationRestrictions,
+} from './index_patterns';
 
 // TODO: figure out how to replace IndexPatterns in get_inner_angular.
+/*
 export {
   IndexPatternsService,
   IndexPatternsContract,
@@ -38,3 +48,4 @@ export {
   TypeMeta,
   AggregationRestrictions,
 } from './index_patterns';
+*/
