+++ src/plugins/data/public/search/search_source/create_search_source.ts
@@ -18,7 +18,7 @@
  */
 import { migrateLegacyQuery } from '../../../../kibana_legacy/public';
 import { SearchSource, SearchSourceDependencies } from './search_source';
-import { IndexPatternsContract } from '../../index_patterns/index_patterns';
+import { IndexPatternsContract } from '../../index_patterns';
 import { SearchSourceFields } from './types';
 
 /**
